Resolv from comments in PR. moved stuff around
