+++ pkg/tss/constants.go
@@ -2,30 +2,18 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// Package tss provides TPM 1.2/2.0 core functionality and
-// abstraction layer for high-level functions
 package tss
 
-// Locality type
-type Locality byte
+import (
+	"crypto"
+	"fmt"
+
+	"github.com/google/go-tpm/tpm2"
+)
 
 // TPMVersion is used to configure a preference in
 // which TPM to use, if multiple are available.
 type TPMVersion uint8
-
-// TPMInterface indicates how the client communicates
-// with the TPM.
-type TPMInterface uint8
-
-// Values of locality
-// Note: Localities are summable
-const (
-	LocZero Locality = 1 << iota
-	LocOne
-	LocTwo
-	LocThree
-	LocFour
-)
 
 // TPM versions
 const (
@@ -34,9 +22,54 @@ const (
 	TPMVersion20
 )
 
+// TPMInterface indicates how the client communicates
+// with the TPM.
+type TPMInterface uint8
+
 // TPM interfaces
 const (
 	TPMInterfaceDirect TPMInterface = iota
 	TPMInterfaceKernelManaged
 	TPMInterfaceDaemonManaged
 )
+
+// HashAlg is the TPM hash algorithm id
+type HashAlg uint8
+
+var (
+	// HashSHA1 is the TPM 1.2 identifier for SHA1
+	HashSHA1 = HashAlg(tpm2.AlgSHA1)
+	// HashSHA256 is the TPM 2.0 identifier for SHA256
+	HashSHA256 = HashAlg(tpm2.AlgSHA256)
+)
+
+func (a HashAlg) cryptoHash() crypto.Hash {
+	switch a {
+	case HashSHA1:
+		return crypto.SHA1
+	case HashSHA256:
+		return crypto.SHA256
+	}
+	return 0
+}
+
+func (a HashAlg) goTPMAlg() tpm2.Algorithm {
+	switch a {
+	case HashSHA1:
+		return tpm2.AlgSHA1
+	case HashSHA256:
+		return tpm2.AlgSHA256
+	}
+	return 0
+}
+
+// String returns a human-friendly representation of the hash algorithm.
+func (a HashAlg) String() string {
+	switch a {
+	case HashSHA1:
+		return "SHA1"
+	case HashSHA256:
+		return "SHA256"
+	}
+	return fmt.Sprintf("HashAlg<%d>", int(a))
+}
